macros: adjust rust flags and cache directory
The cache directory is meant to optimize subsequent rebuilds of the
crates in the "os" package. We weren't seeing much benefit for two
reasons.

First, when building the apiclient and the migrations with the musl
target, we were clobbering the cache for the gnu target, which then
needed to be rebuilt the next time around.

Second, in 8f9306cf we left it up to cargo to decide which flags to
pass to rustc for LTO. It passed `-Clinker-plugin-lto` for apiclient
but not for the migrations, so the cache was immediately invalidated
and needed to be rebuilt.

We solve these problems by using a separate cache directory for the
gnu and musl targets, and by specifying the rustc flags we expect.

Signed-off-by: Ben Cressey <bcressey@amazon.com>