macros: adjust compression for debuginfo and rpms
We spend a significant amount of time compressing debuginfo for
limited gain, since the eventual use case for these symbols is to
archive them for retrieval from a symbol server. Turning it off
improves overall build times.

For compressing the RPMs there are two factors at work. Disabling any
compression speeds up the build, but the larger artifacts take longer
to copy out. Using zstd strikes a reasonable balance between the time
to compress and the time to copy.

Signed-off-by: Ben Cressey <bcressey@amazon.com>