add vendor targets to Rust
Previously, our use of the host platform target was overloaded: it
was used to compile any host binaries needed to build target crates,
and it was also used to build target crates if the host architecture
and the target architecture were the same.

This worked but was brittle. For example, if the version of glibc in
the target's sysroot was newer than the host's, then host binaries
would link correctly but fail to execute with symbol version errors.

Adding vendor-specific targets to the Rust build solves the problem,
since we can leave the host platform target settings untouched, and
specify our cross-compiler toolchain only for our own targets.

Signed-off-by: Ben Cressey <bcressey@amazon.com>