Add license-tool
There are cases where multiple licenses could apply to a package in the
main bottlerocket repo, depending on who will distribute the
software. For such packages, the licensing information has to be
provided at build-time, but can't be hard-coded in the RPM recipes.

`license-tool` can be used to retrieve the license information for a package
with multiple distribution licenses. It uses a toml configuration file
with the licenses that apply to a set of packages. This allows
Bottlerocket's users to build the OS packages with the license that
better applies to them.

Signed-off-by: Arnaldo Garcia Rincon <agarrcia@amazon.com>