Add bootconfig CLI tool
This adds the `bootconfig` CLI tool to the SDK to enable users to use it
without needing to compile it themselves.  Though the tool doesn't get
updated often, it does pull in headers from the kernel proper.  The tool
provided here matches up with the Bottlerocket kernel version at this
time (5.10.109).