license-scan: allow multiple semver crates
With the recent change to handle license clarification with multiple
crates, we introduced a second, newer semver dependency. This is find
and expected, but now causes an issue when building the SDK. During the
SDK build, when `cargo-deny` is run against the license-scan tool, it
now complains about having these multiple versions.

This adds a `skip` setting to the `deny.toml` file to instruct
`cargo-deny` to ignore the older version until we are able to move to a
single crate.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>