go: add helpers for aws-lc syso files
The upstream project's `build-goboring.sh` script contains logic to
generate the self-contained syso files expected by the Go toolchain.
However, it also expects to use the host's toolchain directly, by way
of `clang`, `ld`, and `objcopy` invocations. This complicates the use
of the cross-compiling toolchains needed for Bottlerocket.

Factor out the reusable parts of the script in a way that allows for
custom toolchain use. `parse-functions.go` performs the same function
as the embedded awk scripts, while `umod-amd64.s` and `umod-arm64.c`
are extracted with no changes.

Signed-off-by: DJ Edwards <djedward@amazon.com>
[bcressey: incorporated into SDK]
Signed-off-by: Ben Cressey <bcressey@amazon.com>