build: update Makefile and scripts for unified SDK
The ARCH parameter is no longer used since the SDK is now built for
all supported architectures. Stop passing it to Docker and do not add
it to the image tag.

The toolchain container is no longer built since the SDK has all the
files it previously contained. Stop attempting to build and publish
it.

Signed-off-by: Ben Cressey <bcressey@amazon.com>