macros: set linker flag to pack relative relocations
For shared libraries and position-independent executables, the linker
will generate relative relocations. Newer linkers and dynamic loaders
support the DT_RELR ELF section which allows these relocations to be
stored in a compressed format.

The relevant timeline for Bottlerocket:
* 2022-02: supported by binutils 2.38 for x86_64
* 2022-04: supported by glibc 2.36
* 2023-05: supported by musl 1.2.4
* 2023-11: supported by elfutils 0.190
* 2024-08: supported by binutils 2.43 for aarch64

Now that it's supported by all dependencies for both architectures,
it's time to turn it on.

Disk savings for `/usr/bin` on the `aws-k8s-1.32` variant:
* aarch64: 416 MiB before -> 382 MiB after (~ 8%)
* x86_64:  490 MiB before -> 454 MiB after (~ 7%)

Signed-off-by: Ben Cressey <bcressey@amazon.com>