macros: add set_cross_go_flags_static macro
With the integration of AWS-LC, static builds become more challenging
because `CGO_ENABLED=1` must be set in order for calls into the syso
to work. This is compounded by the difficulties in getting a static
binary when using glibc: AWS-LC has functions that resolve DNS names,
which depends on glibc's NSS, and NSS really does not play well with
static linking.

To avoid rediscovering these pitfalls and reinventing the necessary
solution every time a static Go binary is needed, add a new macro for
setting up the cross-compiling environment for static binaries, which
uses the musl target instead of the glibc one.

Signed-off-by: Ben Cressey <bcressey@amazon.com>