go: remove GOROOT_FINAL
GOROOT_FINAL is not supported in Go 1.23. the bottlerocket-sdk wasn't
using it anyway :) In the sdk-final stage of the Dockerfile, the Go sources
and build artifacts for respective versions are copied over directly
from /home/builder/sdk-go directory in a Go version's respective
directory.

This commit also disables a shellcheck that thinks the  GO111MODULE
environment variable in helpers/go/build-go.sh is unused.

Signed-off-by: Gavin Inglis <giinglis@amazon.com>