macros, build: override find-debuginfo location
This fixes a regression where `find-debuginfo` now prefers `objcopy`
in `/usr/bin` rather than the target-specific symlink placed ahead of
it in $PATH. That happens because `find-debuginfo` now adds its own
parent directory to $PATH in front of everything else.

Work around this by adding a `find-debuginfo` symlink to the debug
helper directory, and override the macro so that `find-debuginfo` is
called from that directory.

Also add a `readelf` symlink, since `find-debuginfo` calls it too.

Signed-off-by: Ben Cressey <bcressey@amazon.com>