build: update base image to Fedora 41
Deal with the behavior change in `rpm2cpio` from RPM 4.20: files are
now prefixed with "./", and when extracting a subset of files, this
prefix must be provided or no match will be found.

Undefine %_auto_set_build_flags in the default macros, since it's no
longer trivial to do so downstream, and they clash with the required
build flags for cross-compiling.

efitools is not used downstream and is no longer available in Fedora,
so drop it.

The OpenSSL ENGINE API headers are no longer installed by default, so
add them back. They can be dropped again when we update the AWS SDK
for C++ with a new-enough version of s2n, which is not yet available.

The gold linker is no longer installed by default, so add a Go patch
so that it's not preferred on aarch64.

Signed-off-by: Ben Cressey <bcressey@amazon.com>