remove cargo-make, docker-cli, and ORAS
"cargo make" was added to the SDK during early Twoliter development,
but since Twoliter has to arrange for the right SDK to be pulled, it
can't easily depend on anything in the SDK.

The CLIs for Docker and ORAS were added for the same purpose and face
similar integration challenges. Twoliter has evolved to bundle krane
instead, and is unlikely to ever use these tools.

Signed-off-by: Ben Cressey <bcressey@amazon.com>