macros: add %_cross_buildsrcdir
Twoliter mounts sources for the current workspace into the SDK build
environment at `/home/builder/rpmbuild/BUILD/sources`, which packages
can refer to via the %_builddir macro - e.g. "%{_builddir}/sources" -
to avoid hard-coding assumptions about the SDK environment, such as
the user account that executes the build.

In RPM 4.20, the %_builddir macro now points to a package-specific
directory rather than the top-level build directory, which breaks
this approach.

Define %_cross_buildsrcdir as a replacement for %_builddir which will
always point to the path in the SDK where Twoliter mounts the sources
from the current workspace.

Signed-off-by: Ben Cressey <bcressey@amazon.com>