build: use clang from Rust bootstrap
Rather than using the most recent version of clang available in the
selected Fedora release, source it from the Rust build.

That provides an indirect form of pinning and control over what ships
in the SDK, which is useful if BPF programs and other binaries might
be compiled with clang.

Signed-off-by: Ben Cressey <bcressey@amazon.com>