rust: set panic=abort for custom targets
Set the panic strategy to "abort" for custom targets, since we expect
that Rust crates for Bottlerocket will be written to avoid panics,
and don't want to pay the cost in binary size to support unwinding.

This lets us avoid specifying the "abort" behavior in Cargo.toml at
the crate or workspace level, so that release builds using the normal
targets on developer systems will default to "unwind", which can be
helpful for debugging.

Signed-off-by: Ben Cressey <bcressey@amazon.com>