build: serialize aws-lc builds inside a Go version
The delocate phase of the aws-lc build can use up to 25 GiB of RAM,
and running eight of them in parallel can lead to OOM even for build
hosts with lots of memory.

Force each stage within a Go version to depend on the stage for the
previous build, so that at most two versions will build concurrently.

Signed-off-by: Ben Cressey <bcressey@amazon.com>