build: add support for arch-specific build flags
Set up scaffolding to pass arch-specific flags in all the places they
may be useful.

Macros will cover most cases, but for various control flow hardening
flags, it's important that the C libraries in the SDK are built in
the same way as the downstream libraries will be, or else the linker
will not automatically annotate the binaries to indicate support.

Signed-off-by: Ben Cressey <bcressey@amazon.com>