macros: enable minimal debuginfo for all builds
Ensure that minimal debuginfo covering functions and line numbers is
included in all binaries. This makes it straightforward to use tools
like `perf` to profile a running system and graph function calls.

Disable the various macros that extract and compress debuginfo, since
that causes the minimal debuginfo to be extracted and packaged inside
separate RPMs.

For downstream variants using EROFS to compress the root filesystem,
the overall cost is around 5% increased disk utilization, or around
15 MiB for a 300 MiB install. That's about the same as adding another
Go program to the default install.

Signed-off-by: Ben Cressey <bcressey@amazon.com>