go: hardcode go build flag -buildvcs=false
On Go-1.24+, go build command now sets the main module's version in the
compiled binary based on VCS tag and/or commit. A +dirty suffix will be
appended if there are uncommitted changes like patches. So we use the
`-buildvcs=false` flag to omit version control information from the binary.

Signed-off-by: Piyush Jena <jepiyush@amazon.com>